gcp/requirements.txt
Pillow>=10.0.0
google-cloud-storage>=2.14.0
python-dotenv>=1.0.0
functions-framework==3.4.0